% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{calc_entropy}
\alias{calc_entropy}
\title{Calculate Shannon Entropy for a Gene Expression Vector}
\usage{
calc_entropy(data_series)
}
\arguments{
\item{data_series}{Numeric vector of gene expression data}

\item{left}{Numeric, left boundary value for histogram bins}

\item{right}{Numeric, right boundary value for histogram bins}

\item{step}{Numeric, interval for histogram bins}
}
\value{
Numeric value representing the Shannon entropy
}
\description{
Calculate Shannon Entropy for a Gene Expression Vector
}
