% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sihouette_score.R
\name{clusterTH_silhouette_score}
\alias{clusterTH_silhouette_score}
\title{Calculate Silhouette Scores for a Seurat Object}
\usage{
clusterTH_silhouette_score(seurat_obj, cluster_col)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{cluster_col}{A column name in the metadata specifying clusters.}
}
\value{
A data.frame containing the silhouette scores for PCA, UMAP, and t-SNE embeddings.
}
\description{
This function calculates silhouette scores for each cell in a Seurat object based on the clusters specified in a metadata column. It adds the silhouette scores to the metadata for PCA, UMAP, and t-SNE embeddings.
}
\examples{
library(Seurat)
# Assuming seurat_obj is a Seurat object and "Cell_line" is a metadata column
scores <- clusterTH_silhouette_score(seurat_obj, "Cell_line")
head(scores)
}
