% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{get_default_ref}
\alias{get_default_ref}
\title{Calculate the Default Reference Vector (Mean Expression) with optional log2 transformation}
\usage{
get_default_ref(data_matrix)
}
\arguments{
\item{data_matrix}{A matrix containing single-cell RNA-seq data}
}
\value{
A numeric vector of mean expression values across all cells
}
\description{
Calculate the Default Reference Vector (Mean Expression) with optional log2 transformation
}
