% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{RCSA}
\alias{RCSA}
\title{RCSA to Find the Intrinsic Reference Cell}
\usage{
RCSA(data_matrix, components = 2)
}
\arguments{
\item{components}{Integer, number of Gaussian components in the model}

\item{seurat_obj}{A Seurat object}
}
\value{
Numeric vector, refined reference vector after RCSA
}
\description{
RCSA to Find the Intrinsic Reference Cell
}
