% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scatter_plot_seurat}
\alias{scatter_plot_seurat}
\title{Scatter Plot from Seurat Object Metadata}
\usage{
scatter_plot_seurat(seurat_obj, x_col, y_col, cluster_col)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{x_col}{The column name to be used for the x-axis.}

\item{y_col}{The column name to be used for the y-axis.}

\item{cluster_col}{The column name to be used for coloring the dots.}
}
\description{
This function generates a scatter plot using two columns from the metadata of a Seurat object and colors the dots based on a specified column of clusters.
}
\examples{
library(Seurat)
# Assuming seurat_obj is a Seurat object with appropriate metadata
scatter_plot_seurat(seurat_obj, "UMAP_1", "UMAP_2", "Cell_Type")
}
