% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Generate Shannon Entropy for intrinsic transcription heterogeneity in a cell}
\usage{
shannon_entropy(
  seurat_obj,
  ref_vec = NULL,
  method = "predefined",
  check_log2 = TRUE,
  scale = FALSE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{ref_vec}{Optional numeric vector, reference cell expression vector}

\item{method}{Character string, method for generating reference vector. It must be either "predefined" or "RCSA".}

\item{check_log2}{Logical, whether to check and perform log2 transformation if necessary}

\item{scale}{Normalize the entropy values to 0-1 range}
}
\value{
Numeric vector, Shannon entropy values relative to the reference
}
\description{
This function was amended from the scEntropy python codes (https://github.com/jzlei/scEntropy) provided by https://www.worldscientific.com/doi/abs/10.1142/S1793048020500010
}
